+++ src/database/operations/authorization/index.ts
@@ -5,3 +5,4 @@ export { hasFunderPermission } from './hasFunderPermission';
 export { hasOpportunityPermission } from './hasOpportunityPermission';
 export { hasProposalFieldValuePermission } from './hasProposalFieldValuePermission';
 export { hasProposalPermission } from './hasProposalPermission';
+export { hasSourcePermission } from './hasSourcePermission';